+++ doc/source/contributing.rst
@@ -775,13 +775,13 @@ Tests that we have ``parametrized`` are now accessible via the test name, for ex
    test_cool_feature.py::test_dtypes[int8] PASSED
    test_cool_feature.py::test_series[int8] PASSED
 
-Transitioning to ``hypothesis``
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+Using ``hypothesis``
+~~~~~~~~~~~~~~~~~~~~
 With the transition to pytest, things have become easier for testing by having reduced boilerplate for test cases and also by utilizing pytest's features like parametizing, skipping and marking test cases.
 
 However, one has to still come up with input data examples which can be tested against the functionality. There is always a possibility to skip testing an example which could have failed the test case.
 
-Imagine if some framework could generate random input examples based on the property/specification of the function being tested. That is exactly what hypothesis does by generating the input data based on some set of specifications provided by the user.
+Hypothesis is a python package which helps in overcoming this issue by generating the input data based on some set of specifications provided by the user.
 e.g suppose we have to test python's sum function for a list of int.
 
 Here is a sample test case using pytest:
@@ -841,6 +841,8 @@ output of test cases:
     ========================== 1 passed in 0.33 seconds ===========================
 
 The main difference in above example is use of a decorator "@given(st.lists(st.integers()))" which if applied to test case function, generates some random list of int, which is then assigned to parameter of test case.
+Above example clearly helps in adding more coverage for our test functions.
+
 For more information about hypothesis or in general about property based testing, check below links:
 
 - https://hypothesis.readthedocs.io/en/latest/quickstart.html
